requirements.txt
django>=1.11.23,<3
djangorestframework>=3.9.0,<4
django-filter>=1.1.0,<2
